refactor(auth): rename folder to metadata
