+++ src/auth/auth.controller.ts
@@ -28,9 +28,9 @@ import { User } from "../user/entity/user.entity";
 import { UserService } from "../user/user.service";
 import { isArrayEmpty } from "../utils/validator";
 import { AuthService } from "./auth.service";
-import { CheckOwner } from "./decorator/owner";
-import { PurposeOneTimeToken } from "./decorator/purposeOneTimeToken";
-import { SecurityTokenGuardOptions } from "./decorator/securityTokenGuardOption";
+import { CheckOwner } from "./metadata/owner";
+import { PurposeOneTimeToken } from "./metadata/purposeOneTimeToken";
+import { SecurityTokenGuardOptions } from "./metadata/securityTokenGuardOption";
 import { CreateOneTimeTokenDTO } from "./dto/request/createOneTimeTokenDTO";
 import { requestVerificationDTO } from "./dto/request/requestVerificationDTO";
 import { SignInResponse } from "./dto/response/signInResponse";
