+++ src/auth/metadata/securityTokenGuardOption.ts
