+++ src/user/user.controller.ts
@@ -20,9 +20,9 @@ import { isEmail, isEmpty, isNotEmpty } from "class-validator";
 import { QueryRunner } from "typeorm";
 import { ServiceException } from "../_common/filter/exception/service/serviceException";
 import { AuthService } from "../auth/auth.service";
-import { CheckOwner } from "../auth/decorator/owner";
-import { PurposeOneTimeToken } from "../auth/decorator/purposeOneTimeToken";
-import { SecurityTokenGuardOptions } from "../auth/decorator/securityTokenGuardOption";
+import { CheckOwner } from "../auth/metadata/owner";
+import { PurposeOneTimeToken } from "../auth/metadata/purposeOneTimeToken";
+import { SecurityTokenGuardOptions } from "../auth/metadata/securityTokenGuardOption";
 import { SecurityTokenGuard } from "../auth/guard/authentication/securityToken.guard";
 import { TempUserGuard } from "../auth/guard/authentication/tempUser.guard";
 import { TokenAuthGuard } from "../auth/guard/authentication/tokenAuth.guard";
